# api/openapi/auth.yml
openapi: 3.0.1
info:
  title: Mainflux authentication service
  description: HTTP API for managing platform API keys.
  version: "1.0.0"
paths:
...
  /groups/{groupId}:
    get:
      summary: Gets group info.
      description: |
        Gets info on a group specified by id.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        '200':
          $ref: "#/components/responses/GroupRes"
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
    put:
      summary: Updates group data.
      description: |
        Updates Name, Description or Metadata of a group.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        $ref: "#/components/requestBodies/GroupUpdateReq"
      responses:
        '200':
          description: Group updated.
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
    delete:
      summary: Deletes group.
      description: |
        Deletes group. If group is parent and descendant groups do not have any members
        child groups will be deleted. Group cannot be deleted if has members or if
        any descendant group has members.
      tags:
        - auth
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/Level"
        - $ref: "#/components/parameters/Metadata"
        - $ref: "#/components/parameters/Tree"
      responses:
        '204':
          description: Group removed.
        '400':
          description: Failed due to malformed query parameters.
        '401':
          description: Missing or invalid access token provided.
        '404':
          description: Group does not exist.
        '500':
          $ref: "#/components/responses/ServiceError"
...
components:
  schemas:
...
    GroupUpdateSchema:
      type: object
      properties:
        name:
          type: string
          description: |
            Free-form group name. Group name is unique on the given hierarchy level.
        description:
          type: string
          description: Group description, free form text.
        metadata:
          type: object
          description: Arbitrary, object-encoded group's data.
    GroupResSchema:
      type: object
      properties:
        id:
          type: string
          format: ulid
          description: Unique group identifier generated by the service.
        name:
          type: string
          description: Free-form group name.
        parent_id:
          type: string
          description: Group ID of parent group.
        owner_id:
          type: string
          format: uuid
          description: UUID of user that created the group.
        metadata:
          type: object
          description: Arbitrary, object-encoded group's data.
        level:
          type: integer
          description: Level in hierarchy, distance from the root group.
        path:
          type: string
          description: Hierarchy path, concatenated ids of group ancestors.
        children:
          type: object
          # schema: GroupResSchema
        created_at:
          type: string
          description: Datetime of group creation.
        updated_at:
          type: string
          description: Datetime of last group updated.
      required:
        - id
        - name
        - owner_id
        - description
        - level
        - path
        - created_at
        - updated_at
...
  parameters:
...
    GroupId:
      name: groupId
      description: Group ID.
      in: path
      schema:
        type: string
        format: uuid
      required: true
...
    Level:
      name: level
      description: Level of hierarchy up to which to retrieve groups from given group id.
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 5
      required: false
    Metadata:
      name: metadata
      description: Metadata filter. Filtering is performed matching the parameter with metadata on top level. Parameter is json.
      in: query
      required: false
      schema:
        type: object
        additionalProperties: {}
    Tree:
      name: tree
      description: Specify type of response, JSON array or tree.
      in: query
      required: false
      schema:
        type: boolean
        default: false
  requestBodies:
...
    GroupUpdateReq:
      description: JSON-formatted document describing group create request.
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GroupUpdateSchema"
...
  responses:
    ServiceError:
      description: Unexpected server-side error occurred.
...
    GroupRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GroupResSchema"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        * Users access: "Authorization: Bearer <user_token>"
security:
  - bearerAuth: []